requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
django-celery-beat==2.5.0
drf-yasg==1.21.7
django-allauth==0.61.1
stripe==7.11.0
Pillow==10.2.0
python-dotenv==1.0.1
whitenoise==6.6.0
geopy==2.4.1
qrcode==7.4.2
redis==5.0.1
celery==5.3.6
setuptools==69.0.3
geopy==2.4.1